Refactor of components inside of conversation state handler and organisation of types related to messaging
Closes #31
