+++ app/(dashboard)/conversations/page.tsx
@@ -1,9 +1,7 @@
 //Components
 import MeatballIcon from '@/components/icons/MeatballIcon';
 import PlusIcon from '@/components/icons/PlusIcon';
-import ConversationStateHandler, {
-  ConversationStateHandlerProps,
-} from '@/components/messaging/CoversationStateHandler';
+import ConversationStateHandler from '@/components/messaging/CoversationStateHandler';
 import { createServerComponentClient } from '@supabase/auth-helpers-nextjs';
 import { cookies } from 'next/headers';
 
@@ -28,15 +26,10 @@ const Conversations = async () => {
         </button>
       </div>
       <div className='mt-4'>
-        {allConversations?.length === 0 ? (
-          <p className='p-10 italic'>
-            {' '}
-            You do not have any conversations active{' '}
-          </p>
+        {allConversations ? (
+          <ConversationStateHandler allConversations={allConversations} />
         ) : (
-          <ConversationStateHandler
-            allConversations={allConversations as ConversationStateHandlerProps}
-          />
+          <p> You do not have any conversations active </p>
         )}
       </div>
     </>
