+++ components/messaging/CoversationStateHandler.tsx
@@ -1,48 +1,31 @@
 'use client';
 
-import { useState } from 'react';
-import { ConversationsListProps } from './ConversationsList';
-import ConversationsList from './ConversationsList';
+import { useEffect, useState } from 'react';
+
+//Types
+import { AllConversationsType } from '@/utils/messaging/messagingTypes';
+
+//Components
+import ConversationsList from './ConversationList';
 import OpenConversation from './OpenConversation';
 
-export type ConversationStateHandlerProps = ConversationsListProps[];
+export type ConversationStateHandlerProps = AllConversationsType;
 
 const ConversationStateHandler = ({
   allConversations,
 }: {
   allConversations: ConversationStateHandlerProps;
 }) => {
-  const [openConvo, setOpenConvo] = useState<ConversationsListProps>(
-    allConversations[0]
-  );
+  const [openConvo, setOpenConvo] = useState(allConversations[0]);
 
-  const updateOpenConvo = (i: number) => {
-    setOpenConvo(allConversations[i]);
-    console.log(openConvo);
-  };
+  useEffect(() => {});
 
   return (
     <>
-      {allConversations.map(
-        (conversation: ConversationsListProps, index: number) => (
-          <div key={`${conversation.id}-${index}`}>
-            <button
-              type='button'
-              onClick={() => {
-                updateOpenConvo(index);
-              }}
-            >
-              <ConversationsList
-                id={conversation.conversation_id}
-                joined_at={conversation.joined_at}
-                conversation_id={conversation.conversation_id}
-                user_id={conversation.user_id}
-                conversations={conversation.conversations}
-              />
-            </button>
-          </div>
-        )
-      )}
+      <ConversationsList
+        allConversations={allConversations}
+        setOpenConvo={setOpenConvo}
+      />
       <OpenConversation
         conversation_id={openConvo.conversation_id}
         user_id={openConvo.user_id}
