+++ utils/supabase/types.ts
@@ -53,15 +53,6 @@ export type user_conversation = {
   user_id: string;
 };
 
-export type message = {
-  id: number;
-  is_read: boolean;
-  sender_id: string;
-  created_at: string;
-  message_text: string;
-  conversation_id: number;
-};
-
 // Make all fields in item and profile optional
 export type PartialItem = Partial<item>;
 export type PartialProfile = Partial<profile>;
